+++ packages/kernel/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - dynamic pawn binding
 - Actor tags
 - AM_Spatial & PM Spatial hav up and forward properties
+- Spatial can have a non-spatial parent
 - NavGrid & PathToBehavior
 
 ### Changed